extract micro mode peer connection into module
